src/requirements.txt
fastapi==0.110.1
sqlmodel==0.0.16
pydantic[email]==2.6.4
sqlalchemy==2.0.29
python-dotenv==1.0.1
boto3==1.34.79
asyncpg==0.29.0
redis[hiredis]==5.0.3
uvicorn==0.29.0
pytest==8.1.1
pytest-asyncio==0.23.6
psycopg2-binary==2.9.9
asyncio==3.4.3
httpx==0.27.0
asgi-lifespan==2.1.0
trio==0.25.0